Extracted waste resource handler into separate class
